requirements.txt
langchain-openai==0.1.22
chromadb==0.5.5
openai
python-dotenv==1.0.1
streamlit==1.37.1
pypdf==4.3.1
python-docx==1.1.2
pandas==2.2.2
